requirements.txt
# Requisitos (requirements.txt):
 streamlit>=1.34,<2
 yfinance>=0.2.43
 pandas>=2.2
 numpy>=1.26,<3
 requests>=2.31
 python-dotenv>=1.0.1
 statsmodels>=0.14.1
 beautifulsoup4>=4.12
 lxml>=4.9.3
 plotly>=5.22
google-genai>=0.3.0  